API/requirements.txt
torch==1.12.0
torchvision==0.13.0
torchaudio==0.12.0
opencv-python==4.6.0.66
pandas==1.4.3
PyYAML==6.0
tqdm==4.64.0
matplotlib==3.5.2
ipython==8.4.0
protobuf==3.20.1
psutil==5.9.1
numpy==1.23.1
Pillow==9.2.0
Flask==2.2.1
easyocr==1.5.0
requests==2.28.1
pytesseract==0.3.10